requirements.txt
streamlit
pillow
python-docx
langchain
langchain-core
langchain_google_genai
python-dotenv
langchain-community